Split PDF mapping helpers
