+++ apps/server/vibesensor/adapters/pdf/presentation.py
@@ -1,12 +1,18 @@
-"""Presentation-only strength labeling helpers for PDF/report rendering."""
+"""Presentation-only labeling helpers for PDF/report rendering."""
 
 from __future__ import annotations
 
 import math
+from collections.abc import Callable
 
 from vibesensor.strength_bands import BANDS
 
-__all__ = ["strength_label", "strength_text"]
+__all__ = [
+    "order_label_human",
+    "peak_classification_text",
+    "strength_label",
+    "strength_text",
+]
 
 _isfinite = math.isfinite
 
@@ -26,6 +32,46 @@ _STRENGTH_THRESHOLDS: tuple[tuple[float, str, str, str], ...] = tuple(
     )
     for band in BANDS
 )
+
+_ORDER_LABEL_NAMES_NL: dict[str, str] = {
+    "wheel": "wielorde",
+    "engine": "motororde",
+    "driveshaft": "aandrijfasorde",
+}
+_ORDER_LABEL_NAMES_DEFAULT: dict[str, str] = {
+    "wheel": "wheel order",
+    "engine": "engine order",
+    "driveshaft": "driveshaft order",
+}
+
+_CLASSIFICATION_I18N_KEYS: dict[str, str] = {
+    "patterned": "CLASSIFICATION_PATTERNED",
+    "persistent": "CLASSIFICATION_PERSISTENT",
+    "transient": "CLASSIFICATION_TRANSIENT",
+    "baseline_noise": "CLASSIFICATION_BASELINE_NOISE",
+}
+
+
+def order_label_human(lang: str, label: str) -> str:
+    """Translate a language-neutral order label like ``1x wheel``."""
+    names = _ORDER_LABEL_NAMES_NL if lang == "nl" else _ORDER_LABEL_NAMES_DEFAULT
+    parts = label.strip().split(" ", 1)
+    if len(parts) == 2:
+        multiplier, base = parts
+        localized = names.get(base.lower(), base)
+        return f"{multiplier} {localized}"
+    return label
+
+
+def peak_classification_text(value: object, tr: Callable[..., str]) -> str:
+    """Map a peak classification code to report text."""
+    normalized = str(value or "").strip().lower()
+    i18n_key = _CLASSIFICATION_I18N_KEYS.get(normalized)
+    if i18n_key:
+        return tr(i18n_key)
+    if not normalized:
+        return tr("UNKNOWN")
+    return str(value).replace("_", " ").title()
 
 
 def strength_label(db_value: float | None, *, lang: str = "en") -> tuple[str, str]:
